Moved LearningRule to learning_rules.py
